Add a possible justification for the multiarch specifier change.
This is without the commercial reasons (see the linked PR) to hopefully
make the change more palatable to the angry mob of unknowing readers.

For the record: this is DEFINITELY NOT the full reason behind the
closed-door decision. But this is about as close as we community people
can get in persuading the depressed ourselves, and about all I can
disclose after having multiple private conversations with relevant
parties, without inflicting damage to anyone at Loongson or Loongson as
a whole (we don't really want to openly criticize Loongson for tweaking
a 1-year-old document for primarily commercial reasons, but again it's
arguably fair for a privately-controlled ISA to evolve like this).

See: https://github.com/loongson/LoongArch-Documentation/pull/80